Split default self-repair test modules (#35)
Move the end-to-end repair tests and duration precision regression into
focused submodules so the root default self-repair test file stays under
the repository file-size policy.

Keep the test behaviour unchanged while leaving the root file as the
orchestrator for smaller self-repair test modules.
