+++ src/agent/self_repair/default_tests.rs
@@ -6,7 +6,7 @@ use std::time::Duration;
 use chrono::Utc;
 use rstest::rstest;
 
-use crate::agent::self_repair::default::{DefaultSelfRepair, duration_since};
+use crate::agent::self_repair::default::DefaultSelfRepair;
 use crate::agent::self_repair::{BrokenTool, NativeSelfRepair, RepairResult, StuckJob};
 use crate::context::{ContextManager, JobState};
 use uuid::Uuid;
@@ -310,160 +310,15 @@ async fn repair_broken_tool_returns_manual_without_builder() {
     );
 }
 
-#[cfg(any(test, feature = "self_repair_extras"))]
-#[tokio::test]
-async fn repair_broken_tool_end_to_end_returns_success() {
-    use crate::testing::null_db::CapturingStore;
-
-    use self::helpers::{StubBuilderOutcome, StubSoftwareBuilder};
-
-    let cm = Arc::new(ContextManager::new(10));
-    let store = Arc::new(CapturingStore::new());
-    let store_for_repair: Arc<dyn crate::db::Database> = store.clone();
-
-    let builder = Arc::new(StubSoftwareBuilder(StubBuilderOutcome::BuildSucceeded {
-        is_success: true,
-        error: None,
-        iterations: 1,
-        is_registered: false,
-    }));
-    let builder_for_repair: Arc<dyn crate::tools::SoftwareBuilder> = builder;
-    let tools = Arc::new(crate::tools::ToolRegistry::new());
-
-    let repair = DefaultSelfRepair::new(cm, Duration::from_secs(60), 3)
-        .with_store(store_for_repair)
-        .with_builder(builder_for_repair, tools);
-
-    let broken = BrokenTool {
-        name: "my-tool".to_string(),
-        failure_count: 2,
-        last_error: Some("test error".to_string()),
-        first_failure: Utc::now(),
-        last_failure: Utc::now(),
-        last_build_result: None,
-        repair_attempts: 0,
-    };
-
-    let result = NativeSelfRepair::repair_broken_tool(&repair, &broken)
-        .await
-        .expect("repair_broken_tool should not error");
-
-    assert!(
-        matches!(result, RepairResult::Success { .. }),
-        "expected RepairResult::Success end-to-end, got: {result:?}",
-    );
-    assert_eq!(
-        *store.calls().repaired_tools.lock().await,
-        vec!["my-tool".to_string()],
-        "repair_broken_tool should mark the tool repaired in the store",
-    );
-}
-
-#[cfg(any(test, feature = "self_repair_extras"))]
-#[tokio::test]
-async fn repair_broken_tool_allows_one_concurrent_repair_for_same_tool() {
-    use crate::testing::null_db::CapturingStore;
-
-    use self::helpers::{StubBuilderOutcome, StubSoftwareBuilder};
-
-    let cm = Arc::new(ContextManager::new(10));
-    let store = Arc::new(CapturingStore::new());
-    let store_for_repair: Arc<dyn crate::db::Database> = store.clone();
-
-    let builder = Arc::new(StubSoftwareBuilder(StubBuilderOutcome::BuildSucceeded {
-        is_success: true,
-        error: None,
-        iterations: 1,
-        is_registered: false,
-    }));
-    let builder_for_repair: Arc<dyn crate::tools::SoftwareBuilder> = builder;
-    let tools = Arc::new(crate::tools::ToolRegistry::new());
-
-    let repair = Arc::new(
-        DefaultSelfRepair::new(cm, Duration::from_secs(60), 3)
-            .with_store(store_for_repair)
-            .with_builder(builder_for_repair, tools),
-    );
-
-    let broken = Arc::new(BrokenTool {
-        name: "my-tool".to_string(),
-        failure_count: 2,
-        last_error: Some("test error".to_string()),
-        first_failure: Utc::now(),
-        last_failure: Utc::now(),
-        last_build_result: None,
-        repair_attempts: 0,
-    });
-
-    let first_repair = Arc::clone(&repair);
-    let first_broken = Arc::clone(&broken);
-    let first = tokio::spawn(async move {
-        NativeSelfRepair::repair_broken_tool(first_repair.as_ref(), first_broken.as_ref()).await
-    });
-    let second_repair = Arc::clone(&repair);
-    let second_broken = Arc::clone(&broken);
-    let second = tokio::spawn(async move {
-        NativeSelfRepair::repair_broken_tool(second_repair.as_ref(), second_broken.as_ref()).await
-    });
-
-    let (first, second) = tokio::join!(
-        async { first.await.expect("first repair task should complete") },
-        async { second.await.expect("second repair task should complete") },
-    );
-    let results = [
-        first.expect("first repair_broken_tool call should not error"),
-        second.expect("second repair_broken_tool call should not error"),
-    ];
-
-    assert_eq!(
-        results
-            .iter()
-            .filter(|result| matches!(result, RepairResult::Success { .. }))
-            .count(),
-        1,
-        "exactly one concurrent repair should succeed",
-    );
-    assert!(
-        results.iter().any(|result| matches!(
-            result,
-            RepairResult::Retry { message }
-                if message == "Repair already in progress for 'my-tool'"
-        )),
-        "one concurrent repair should be rejected as already in progress: {results:?}",
-    );
-    assert_eq!(
-        *store.calls().repaired_tools.lock().await,
-        vec!["my-tool".to_string()],
-        "only the claimed repair should mark the tool repaired",
-    );
-}
-
-#[test]
-fn duration_since_millisecond_precision() {
-    use chrono::Duration as ChronoDuration;
-
-    let now = Utc::now();
-    let start = now - ChronoDuration::milliseconds(500);
-    let elapsed = duration_since(now, start);
-
-    // Should be >= 500ms and < 1s (proving millisecond resolution, not second)
-    assert!(
-        elapsed >= Duration::from_millis(500),
-        "Expected >= 500ms, got {:?}",
-        elapsed
-    );
-    assert!(
-        elapsed < Duration::from_secs(1),
-        "Expected < 1s, got {:?}",
-        elapsed
-    );
-}
-
 #[path = "default_tests/attempt_repair_build.rs"]
 mod attempt_repair_build;
 #[path = "default_tests/build_repair_requirement.rs"]
 mod build_repair_requirement;
+#[path = "default_tests/duration_precision.rs"]
+mod duration_precision;
 #[path = "default_tests/handle_build_result.rs"]
 mod handle_build_result;
 #[path = "default_tests/helpers.rs"]
 mod helpers;
+#[path = "default_tests/repair_end_to_end.rs"]
+mod repair_end_to_end;
